Fix for text selection.
Including lightbox.html in the body section leads to text selection issues.

The text on the page becomes unselectible.

I don't know why, but all that "user-select: none" works for global page scope, not only for "lightbox" style.

I tested it for Chrome for Mac.

Fixed in documentation, it now says that lightbox.html should be included in the head section.